orig: Adding the original file to revision control
This is the beginning of a series of checks around fixing [DUID][1]
support on the EdgeRouter platform.  Unfortunately Ubiquiti doesn't
really "open source" their code per se so I'm going to have to cowboy
this a bit.

¯\_(ツ)_/¯

This bug has been outstanding for a year and (seemingly) Ubiquity
can/won't fix it.  You'll find on page 10 of that thread that I finally
started debugging it and seemingly identified the problem.

Adding the original file to revision control so that I can debug this
endianess issue.

[1]: https://community.ubnt.com/t5/EdgeRouter/Comcast-IPv6-issues-when-hwnat-enabled-on-ER-X/m-p/2262797